WIP: removing cycles
