+++ include/virt_wrap/Connection.hpp
@@ -7,6 +7,9 @@
 #include <vector>
 #include <gsl/gsl>
 #include <libvirt/libvirt.h>
+#include "enums/Connection/Decls.hpp"
+#include "enums/Domain/Decls.hpp"
+#include "enums/Domain/SaveRestoreFlag.hpp"
 #include "fwd.hpp"
 #include "utility.hpp"
 
@@ -14,21 +17,8 @@ namespace virt {
 
 unsigned long getVersion();
 
-enum class CredentialType : int {
-    AUTHNAME = VIR_CRED_AUTHNAME,         // Identify to authorize as
-    CNONCE = VIR_CRED_CNONCE,             // Client supplies a nonce
-    ECHOPROMPT = VIR_CRED_ECHOPROMPT,     // Challenge response
-    EXTERNAL = VIR_CRED_EXTERNAL,         // Externally managed credential More may be
-                                          // added - expect the unexpected
-    LANGUAGE = VIR_CRED_LANGUAGE,         // RFC 1766 languages, comma separated
-    NOECHOPROMPT = VIR_CRED_NOECHOPROMPT, // Challenge response
-    PASSPHRASE = VIR_CRED_PASSPHRASE,     // Passphrase secret
-    REALM = VIR_CRED_REALM,               // Authentication realm
-    USERNAME = VIR_CRED_USERNAME,         // Identity to act as
-};
-
 struct ConnectCredential {
-    const CredentialType type;
+    const enums::connection::CredentialType type;
     const gsl::czstring<> prompt;    /* Prompt to show to user */
     const gsl::czstring<> challenge; /* Additional challenge to show */
     const gsl::czstring<> defresult; /* Optional default result */
@@ -49,7 +39,7 @@ template <typename Callback = ConnectAuthCallback> class ConnectionAuth {
 
     inline explicit operator virConnectAuth() noexcept;
 
-    std::vector<CredentialType> cred_types{};
+    std::vector<enums::connection::CredentialType> cred_types{};
     Callback callback;
 };
 
@@ -64,72 +54,12 @@ class Connection {
     virConnectPtr underlying = nullptr;
 
   public:
-    enum class Flags : unsigned {
-        RO = VIR_CONNECT_RO,                 /* A readonly connection */
-        NO_ALIASES = VIR_CONNECT_NO_ALIASES, /* Don't try to resolve URI aliases */
-    };
     struct List {
         struct Domains {
             class Flag;
         };
-        struct Devices {
-            enum class Flags : unsigned {
-                DEFAULT = 0,
-                SYSTEM = VIR_CONNECT_LIST_NODE_DEVICES_CAP_SYSTEM,
-                PCI_DEV = VIR_CONNECT_LIST_NODE_DEVICES_CAP_PCI_DEV,
-                USB_DEV = VIR_CONNECT_LIST_NODE_DEVICES_CAP_USB_DEV,
-                USB_IFACE = VIR_CONNECT_LIST_NODE_DEVICES_CAP_USB_INTERFACE,
-                NETWORK = VIR_CONNECT_LIST_NODE_DEVICES_CAP_NET,
-                SCSI_HOST = VIR_CONNECT_LIST_NODE_DEVICES_CAP_SCSI_HOST,
-                SCSI_TARGET = VIR_CONNECT_LIST_NODE_DEVICES_CAP_SCSI_TARGET,
-                SCSI_DEV = VIR_CONNECT_LIST_NODE_DEVICES_CAP_SCSI,
-                STORAGE = VIR_CONNECT_LIST_NODE_DEVICES_CAP_STORAGE,
-                FC_HOST = VIR_CONNECT_LIST_NODE_DEVICES_CAP_FC_HOST,
-                VPORTS = VIR_CONNECT_LIST_NODE_DEVICES_CAP_VPORTS,
-                SCSI_GENERIC = VIR_CONNECT_LIST_NODE_DEVICES_CAP_SCSI_GENERIC,
-                DRM = VIR_CONNECT_LIST_NODE_DEVICES_CAP_DRM,
-                MEDIATED_CAPABLE = VIR_CONNECT_LIST_NODE_DEVICES_CAP_MDEV_TYPES,
-                MEDIATED = VIR_CONNECT_LIST_NODE_DEVICES_CAP_MDEV,
-                CCW = VIR_CONNECT_LIST_NODE_DEVICES_CAP_CCW_DEV,
-            };
-        };
-        struct Networks {
-            enum class Flag : unsigned {
-                DEFAULT = 0,
-
-                INACTIVE = VIR_CONNECT_LIST_NETWORKS_INACTIVE,
-                ACTIVE = VIR_CONNECT_LIST_NETWORKS_ACTIVE,
-
-                PERSISTENT = VIR_CONNECT_LIST_NETWORKS_PERSISTENT,
-                TRANSIENT = VIR_CONNECT_LIST_NETWORKS_TRANSIENT,
-
-                AUTOSTART = VIR_CONNECT_LIST_NETWORKS_AUTOSTART,
-                NO_AUTOSTART = VIR_CONNECT_LIST_NETWORKS_NO_AUTOSTART,
-            };
-        };
         struct StoragePool {
             class Flag;
-        };
-    };
-    struct GetAllDomains {
-        struct Stats {
-            enum class Flags : unsigned {
-                ACTIVE = VIR_CONNECT_GET_ALL_DOMAINS_STATS_ACTIVE,
-                INACTIVE = VIR_CONNECT_GET_ALL_DOMAINS_STATS_INACTIVE,
-
-                PERSISTENT = VIR_CONNECT_LIST_DOMAINS_PERSISTENT,
-                TRANSIENT = VIR_CONNECT_LIST_DOMAINS_TRANSIENT,
-
-                RUNNING = VIR_CONNECT_LIST_DOMAINS_RUNNING,
-                PAUSED = VIR_CONNECT_GET_ALL_DOMAINS_STATS_PAUSED,
-                SHUTOFF = VIR_CONNECT_LIST_DOMAINS_SHUTOFF,
-                OTHER = VIR_CONNECT_GET_ALL_DOMAINS_STATS_OTHER,
-
-                NOWAIT = VIR_CONNECT_GET_ALL_DOMAINS_STATS_NOWAIT,               // report statistics that can be obtained immediately without any
-                                                                                 // blocking
-                BACKING = VIR_CONNECT_GET_ALL_DOMAINS_STATS_BACKING,             // include backing chain for block stats
-                ENFORCE_STATS = VIR_CONNECT_GET_ALL_DOMAINS_STATS_ENFORCE_STATS, // enforce requested stats
-            };
         };
     };
 
@@ -140,7 +70,7 @@ class Connection {
     inline explicit Connection(gsl::czstring<> name, bool rd_only = false) noexcept;
 
     template <typename Callback = virConnectAuthCallbackPtr>
-    inline Connection(gsl::czstring<> name, ConnectionAuth<Callback>& auth, Flags flags) noexcept;
+    inline Connection(gsl::czstring<> name, ConnectionAuth<Callback>& auth, enums::connection::Flags flags) noexcept;
 
     inline Connection(const Connection& conn) noexcept = default;
 
@@ -203,8 +133,8 @@ class Connection {
 
     [[nodiscard]] inline auto listAllDomains(List::Domains::Flag flags) const -> std::vector<Domain>;
 
-    [[nodiscard]] inline auto getAllDomainStats(Domain::Stats::Types stats, Connection::GetAllDomains::Stats::Flags flags)
-        -> std::vector<Domain::Stats::Record>;
+    [[nodiscard]] inline auto getAllDomainStats(enums::domain::stats::Types stats, enums::connection::get_all_domains::stats::Flags flags)
+        -> std::vector<enums::domain::stats::Record>;
 
     [[nodiscard]] inline Domain domainLookupByID(int) const noexcept;
