+++ include/virt_wrap/Domain.hpp
@@ -11,7 +11,8 @@
 #include <vector>
 #include <gsl/gsl>
 #include "../cexpr_algs.hpp"
-#include <libvirt/libvirt-domain.h>
+#include "enums/Connection/Decls.hpp"
+#include "enums/Domain/Decls.hpp"
 #include "CpuMap.hpp"
 #include "GFlags.hpp"
 #include "fwd.hpp"
@@ -98,20 +99,7 @@ class Domain {
 
   public:
     using Info = virDomainInfo;
-    struct Stats {
-        enum class Types {
-            STATE = VIR_DOMAIN_STATS_STATE,         // domain state
-            CPU_TOTAL = VIR_DOMAIN_STATS_CPU_TOTAL, // domain CPU info
-            BALLOON = VIR_DOMAIN_STATS_BALLOON,     // domain balloon info
-            VCPU = VIR_DOMAIN_STATS_VCPU,           // domain virtual CPU info
-            INTERFACE = VIR_DOMAIN_STATS_INTERFACE, // domain interfaces info
-            BLOCK = VIR_DOMAIN_STATS_BLOCK,         // domain block info
-            PERF = VIR_DOMAIN_STATS_PERF,           // domain perf event info
-            IOTHREAD = VIR_DOMAIN_STATS_IOTHREAD,   // iothread poll info
-        };
 
-        class Record;
-    };
     class BlockCommitFlag;
     class BlockCopyFlag;
     class BlockJobAbortFlag;
@@ -123,57 +111,10 @@ class Domain {
     class ChannelFlag;
     class ConsoleFlag;
 
-    struct CoreDump {
-        enum class Flag {
-            CRASH = VIR_DUMP_CRASH,               /* crash after dump */
-            LIVE = VIR_DUMP_LIVE,                 /* live dump */
-            BYPASS_CACHE = VIR_DUMP_BYPASS_CACHE, /* avoid file system cache pollution */
-            RESET = VIR_DUMP_RESET,               /* reset domain after dump finishes */
-            MEMORY_ONLY = VIR_DUMP_MEMORY_ONLY,   /* use dump-guest-memory */
-        };
-        enum class Format {
-            RAW = VIR_DOMAIN_CORE_DUMP_FORMAT_RAW,                   /* dump guest memory in raw format */
-            KDUMP_ZLIB = VIR_DOMAIN_CORE_DUMP_FORMAT_KDUMP_ZLIB,     /* kdump-compressed format, with zlib compression */
-            KDUMP_LZO = VIR_DOMAIN_CORE_DUMP_FORMAT_KDUMP_LZO,       /* kdump-compressed format, with lzo compression */
-            KDUMP_SNAPPY = VIR_DOMAIN_CORE_DUMP_FORMAT_KDUMP_SNAPPY, /* kdump-compressed format, with snappy compression */
-        };
-    };
-
     class CreateFlag;
     class DestroyFlag;
     class DeviceModifyFlag;
     class GetJobStatsFlag;
-
-    enum class GetAllDomainStatsFlag : unsigned {
-        ACTIVE = VIR_CONNECT_GET_ALL_DOMAINS_STATS_ACTIVE,
-        INACTIVE = VIR_CONNECT_GET_ALL_DOMAINS_STATS_INACTIVE,
-
-        PERSISTENT = VIR_CONNECT_GET_ALL_DOMAINS_STATS_PERSISTENT,
-        TRANSIENT = VIR_CONNECT_GET_ALL_DOMAINS_STATS_TRANSIENT,
-
-        RUNNING = VIR_CONNECT_GET_ALL_DOMAINS_STATS_RUNNING,
-        PAUSED = VIR_CONNECT_GET_ALL_DOMAINS_STATS_PAUSED,
-        SHUTOFF = VIR_CONNECT_GET_ALL_DOMAINS_STATS_SHUTOFF,
-        OTHER = VIR_CONNECT_GET_ALL_DOMAINS_STATS_OTHER,
-
-        NOWAIT = VIR_CONNECT_GET_ALL_DOMAINS_STATS_NOWAIT,       /* report statistics that can be obtained
-                                                                immediately without any blocking */
-        BACKING = VIR_CONNECT_GET_ALL_DOMAINS_STATS_BACKING,     /* include backing chain for block stats */
-        STATS = VIR_CONNECT_GET_ALL_DOMAINS_STATS_ENFORCE_STATS, /* enforce requested stats */
-    };
-    enum class InterfaceAddressesSource : unsigned {
-        LEASE = VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_LEASE, /* Parse DHCP lease file */
-        AGENT = VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_AGENT, /* Query qemu guest agent */
-        ARP = VIR_DOMAIN_INTERFACE_ADDRESSES_SRC_ARP,     /* Query ARP tables */
-    };
-    enum class JobType {
-        NONE = VIR_DOMAIN_JOB_NONE,           /* No job is active */
-        BOUNDED = VIR_DOMAIN_JOB_BOUNDED,     /* Job with a finite completion time */
-        UNBOUNDED = VIR_DOMAIN_JOB_UNBOUNDED, /* Job without a finite completion time */
-        COMPLETED = VIR_DOMAIN_JOB_COMPLETED, /* Job has finished, but isn't cleaned up */
-        FAILED = VIR_DOMAIN_JOB_FAILED,       /* Job hit error, but isn't cleaned up */
-        CANCELLED = VIR_DOMAIN_JOB_CANCELLED, /* Job was aborted, but isn't cleaned up */
-    };
 
     class KeycodeSet;
     class Lifecycle;
@@ -181,171 +122,17 @@ class Domain {
     class MemoryModFlag;
     class OpenGraphicsFlag;
 
-    enum class SaveImageXMLFlag : unsigned {
-        SECURE = VIR_DOMAIN_XML_SECURE, /* dump security sensitive information too */
-    };
+    class StatsRecord;
 
-    class SaveRestoreFlag;
     class ShutdownFlag;
 
-    struct StateReason {
-        enum class NoState {
-            UNKNOWN = VIR_DOMAIN_NOSTATE_UNKNOWN, /* the reason is unknown */
-        };
-        enum class Running {
-            UNKNOWN = VIR_DOMAIN_RUNNING_UNKNOWN,                       /* the reason is unknown */
-            BOOTED = VIR_DOMAIN_RUNNING_BOOTED,                         /* normal startup from boot */
-            MIGRATED = VIR_DOMAIN_RUNNING_MIGRATED,                     /* migrated from another host */
-            RESTORED = VIR_DOMAIN_RUNNING_RESTORED,                     /* restored from a state file */
-            FROM_SNAPSHOT = VIR_DOMAIN_RUNNING_FROM_SNAPSHOT,           /* restored from snapshot */
-            UNPAUSED = VIR_DOMAIN_RUNNING_UNPAUSED,                     /* returned from paused state */
-            MIGRATION_CANCELED = VIR_DOMAIN_RUNNING_MIGRATION_CANCELED, /* returned from migration */
-            SAVE_CANCELED = VIR_DOMAIN_RUNNING_SAVE_CANCELED,           /* returned from failed save process */
-            WAKEUP = VIR_DOMAIN_RUNNING_WAKEUP,                         /* returned from pmsuspended due to wakeup event */
-            CRASHED = VIR_DOMAIN_RUNNING_CRASHED,                       /* resumed from crashed */
-            POSTCOPY = VIR_DOMAIN_RUNNING_POSTCOPY,                     /* running in post-copy migration mode */
-        };
-        enum class Blocked {
-            UNKNOWN = VIR_DOMAIN_BLOCKED_UNKNOWN, /* the reason is unknown */
-        };
-        enum class Paused {
-            UNKNOWN = VIR_DOMAIN_PAUSED_UNKNOWN,                 /* the reason is unknown */
-            USER = VIR_DOMAIN_PAUSED_USER,                       /* paused on user request */
-            MIGRATION = VIR_DOMAIN_PAUSED_MIGRATION,             /* paused for offline migration */
-            SAVE = VIR_DOMAIN_PAUSED_SAVE,                       /* paused for save */
-            DUMP = VIR_DOMAIN_PAUSED_DUMP,                       /* paused for offline core dump */
-            IOERROR = VIR_DOMAIN_PAUSED_IOERROR,                 /* paused due to a disk I/O error */
-            WATCHDOG = VIR_DOMAIN_PAUSED_WATCHDOG,               /* paused due to a watchdog event */
-            FROM_SNAPSHOT = VIR_DOMAIN_PAUSED_FROM_SNAPSHOT,     /* paused after restoring from snapshot */
-            SHUTTING_DOWN = VIR_DOMAIN_PAUSED_SHUTTING_DOWN,     /* paused during shutdown process */
-            SNAPSHOT = VIR_DOMAIN_PAUSED_SNAPSHOT,               /* paused while creating a snapshot */
-            CRASHED = VIR_DOMAIN_PAUSED_CRASHED,                 /* paused due to a guest crash */
-            STARTING_UP = VIR_DOMAIN_PAUSED_STARTING_UP,         /* the domain is being started */
-            POSTCOPY = VIR_DOMAIN_PAUSED_POSTCOPY,               /* paused for post-copy migration */
-            POSTCOPY_FAILED = VIR_DOMAIN_PAUSED_POSTCOPY_FAILED, /* paused after failed post-copy */
-        };
-        enum class Shutdown {
-            UNKNOWN = VIR_DOMAIN_SHUTDOWN_UNKNOWN, /* the reason is unknown */
-            USER = VIR_DOMAIN_SHUTDOWN_USER,       /* shutting down on user request */
-        };
-        enum class Shutoff {
-            UNKNOWN = VIR_DOMAIN_SHUTOFF_UNKNOWN,             /* the reason is unknown */
-            SHUTDOWN = VIR_DOMAIN_SHUTOFF_SHUTDOWN,           /* normal shutdown */
-            DESTROYED = VIR_DOMAIN_SHUTOFF_DESTROYED,         /* forced poweroff */
-            CRASHED = VIR_DOMAIN_SHUTOFF_CRASHED,             /* domain crashed */
-            MIGRATED = VIR_DOMAIN_SHUTOFF_MIGRATED,           /* migrated to another host */
-            SAVED = VIR_DOMAIN_SHUTOFF_SAVED,                 /* saved to a file */
-            FAILED = VIR_DOMAIN_SHUTOFF_FAILED,               /* domain failed to start */
-            FROM_SNAPSHOT = VIR_DOMAIN_SHUTOFF_FROM_SNAPSHOT, /* restored from a snapshot which was taken while domain was shutoff */
-            DAEMON = VIR_DOMAIN_SHUTOFF_DAEMON,               /* daemon decides to kill domain during reconnection processing */
-        };
-        enum class Crashed {
-            UNKNOWN = VIR_DOMAIN_CRASHED_UNKNOWN,   /* crashed for unknown reason */
-            PANICKED = VIR_DOMAIN_CRASHED_PANICKED, /* domain panicked */
-        };
-        enum class PMSuspended {
-            UNKNOWN = VIR_DOMAIN_PMSUSPENDED_UNKNOWN, /* the reason is unknown */
-        };
-    };
+    struct StateReason {};
     struct StateWReason;
-    enum class StatsType {
-        STATE = VIR_DOMAIN_STATS_STATE,         /* return domain state */
-        TOTAL = VIR_DOMAIN_STATS_CPU_TOTAL,     /* return domain CPU info */
-        BALLOON = VIR_DOMAIN_STATS_BALLOON,     /* return domain balloon info */
-        VCPU = VIR_DOMAIN_STATS_VCPU,           /* return domain virtual CPU info */
-        INTERFACE = VIR_DOMAIN_STATS_INTERFACE, /* return domain interfaces info */
-        BLOCK = VIR_DOMAIN_STATS_BLOCK,         /* return domain block info */
-        PERF = VIR_DOMAIN_STATS_PERF,           /* return domain perf event info */
-        IOTHREAD = VIR_DOMAIN_STATS_IOTHREAD,   /* return iothread poll info */
-    };
-
-    enum class MemoryFlag {
-        VIRTUAL = VIR_MEMORY_VIRTUAL,   /* addresses are virtual addresses */
-        PHYSICAL = VIR_MEMORY_PHYSICAL, /* addresses are physical addresses */
-    };
-
-    enum class MemoryStatTag {
-        /* The total amount of data read from swap space (in kB). */
-        SWAP_IN = VIR_DOMAIN_MEMORY_STAT_SWAP_IN,
-        /* The total amount of memory written out to swap space (in kB). */
-        SWAP_OUT = VIR_DOMAIN_MEMORY_STAT_SWAP_OUT,
-
-        /*
-         * Page faults occur when a process makes a valid access to virtual memory
-         * that is not available.  When servicing the page fault, if disk IO is
-         * required, it is considered a major fault.  If not, it is a minor fault.
-         * These are expressed as the number of faults that have occurred.
-         */
-        MAJOR_FAULT = VIR_DOMAIN_MEMORY_STAT_MAJOR_FAULT,
-        MINOR_FAULT = VIR_DOMAIN_MEMORY_STAT_MINOR_FAULT,
-
-        /*
-         * The amount of memory left completely unused by the system.  Memory that
-         * is available but used for reclaimable caches should NOT be reported as
-         * free.  This value is expressed in kB.
-         */
-        UNUSED = VIR_DOMAIN_MEMORY_STAT_UNUSED,
-
-        /*
-         * The total amount of usable memory as seen by the domain.  This value
-         * may be less than the amount of memory assigned to the domain if a
-         * balloon driver is in use or if the guest OS does not initialize all
-         * assigned pages.  This value is expressed in kB.
-         */
-        AVAILABLE = VIR_DOMAIN_MEMORY_STAT_AVAILABLE,
-
-        /* Current balloon value (in KB). */
-        ACTUAL_BALLOON = VIR_DOMAIN_MEMORY_STAT_ACTUAL_BALLOON,
-
-        /* Resident Set Size of the process running the domain. This value
-         * is in kB */
-        RSS = VIR_DOMAIN_MEMORY_STAT_RSS,
-
-        /*
-         * How much the balloon can be inflated without pushing the guest system
-         * to swap, corresponds to 'Available' in /proc/meminfo
-         */
-        USABLE = VIR_DOMAIN_MEMORY_STAT_USABLE,
-
-        /* Timestamp of the last update of statistics, in seconds. */
-        UPDATE = VIR_DOMAIN_MEMORY_STAT_LAST_UPDATE,
-
-        /*
-         * The amount of memory, that can be quickly reclaimed without
-         * additional I/O (in kB). Typically these pages are used for caching files
-         * from disk.
-         */
-        DISK_CACHES = VIR_DOMAIN_MEMORY_STAT_DISK_CACHES,
-
-        /*
-         * The number of statistics supported by this version of the interface.
-         * To add new statistics, add them to the enum and increase this value.
-         */
-        NR = VIR_DOMAIN_MEMORY_STAT_NR,
-    };
 
     class MetadataType;
     class MigrateFlag;
     class ModificationImpactFlag;
     class ProcessSignal;
-
-    enum class SetTimeFlag {
-        SYNC = VIR_DOMAIN_TIME_SYNC, /* Re-sync domain time from domain's RTC */
-    };
-    enum class SetUserPasswordFlag {
-        ENCRYPTED = VIR_DOMAIN_PASSWORD_ENCRYPTED, /* the password is already encrypted */
-    };
-    enum class VCpuFlag {
-        /* See ModificationImpactFlag for these flags.  */
-        CURRENT = VIR_DOMAIN_VCPU_CURRENT,
-        LIVE = VIR_DOMAIN_VCPU_LIVE,
-        CONFIG = VIR_DOMAIN_VCPU_CONFIG,
-
-        /* Additionally, these flags may be bitwise-OR'd in.  */
-        MAXIMUM = VIR_DOMAIN_VCPU_MAXIMUM,           /* Max rather than current count */
-        GUEST = VIR_DOMAIN_VCPU_GUEST,               /* Modify state of the cpu in the guest */
-        HOTPLUGGABLE = VIR_DOMAIN_VCPU_HOTPLUGGABLE, /* Make vcpus added hot(un)pluggable */
-    };
 
     class XmlFlag;
     class State;
@@ -417,9 +204,9 @@ class Domain {
 
     // createWithFiles() // Left out
 
-    bool coreDump(std::filesystem::path to, CoreDump::Flag flags) const noexcept;
+    bool coreDump(std::filesystem::path to, enums::domain::core_dump::Flag flags) const noexcept;
 
-    bool coreDump(std::filesystem::path to, CoreDump::Format format, CoreDump::Flag flags) const noexcept;
+    bool coreDump(std::filesystem::path to, enums::domain::core_dump::Format format, enums::domain::core_dump::Flag flags) const noexcept;
 
     bool delIOThread(unsigned int iothread_id, ModificationImpactFlag flags) noexcept;
 
@@ -501,7 +288,7 @@ class Domain {
 
     [[nodiscard]] auto getNumaParameters(MITPFlags flags) const noexcept;
 
-    [[nodiscard]] int getNumVcpus(VCpuFlag flags) const noexcept;
+    [[nodiscard]] int getNumVcpus(enums::domain::VCpuFlag flags) const noexcept;
 
     [[nodiscard]] auto getSchedulerType() const noexcept -> std::pair<UniqueZstring, int>;
 
@@ -533,7 +320,7 @@ class Domain {
 
     [[nodiscard]] auto getPerfEvents(MITPFlags flags) const noexcept;
 
-    [[nodiscard]] auto getVcpuPinInfo(VCpuFlag flags) -> std::optional<std::vector<unsigned char>>;
+    [[nodiscard]] auto getVcpuPinInfo(enums::domain::VCpuFlag flags) -> std::optional<std::vector<unsigned char>>;
 
     [[nodiscard]] auto getVcpus() const noexcept;
 
@@ -543,9 +330,9 @@ class Domain {
 
     bool injectNMI() noexcept;
 
-    [[nodiscard]] auto interfaceAddressesView(InterfaceAddressesSource source) const noexcept;
+    [[nodiscard]] auto interfaceAddressesView(enums::domain::InterfaceAddressesSource source) const noexcept;
 
-    [[nodiscard]] auto interfaceAddresses(InterfaceAddressesSource source) const -> std::vector<Interface>;
+    [[nodiscard]] auto interfaceAddresses(enums::domain::InterfaceAddressesSource source) const -> std::vector<Interface>;
 
     [[nodiscard]] auto interfaceStats(gsl::czstring<> device) const noexcept -> std::optional<virDomainInterfaceStatsStruct>;
 
@@ -563,13 +350,13 @@ class Domain {
 
     bool managedSaveDefineXML(gsl::czstring<> dxml, SaveRestoreFlag flag) noexcept;
 
-    [[nodiscard]] UniqueZstring managedSaveGetXMLDesc(SaveImageXMLFlag flag) const noexcept;
+    [[nodiscard]] UniqueZstring managedSaveGetXMLDesc(enums::domain::SaveImageXMLFlag flag) const noexcept;
 
-    [[nodiscard]] std::string managedSaveExtractXMLDesc(SaveImageXMLFlag flag) const noexcept;
+    [[nodiscard]] std::string managedSaveExtractXMLDesc(enums::domain::SaveImageXMLFlag flag) const noexcept;
 
     bool managedSaveRemove() noexcept;
 
-    bool memoryPeek(unsigned long long start, gsl::span<unsigned char> buffer, MemoryFlag flag) const noexcept;
+    bool memoryPeek(unsigned long long start, gsl::span<unsigned char> buffer, enums::domain::MemoryFlag flag) const noexcept;
 
     auto memoryStats(unsigned int nr_stats) const noexcept;
 
@@ -672,15 +459,15 @@ class Domain {
 
     bool setMetadata(MetadataType type, gsl::czstring<> metadata, gsl::czstring<> key, gsl::czstring<> uri, ModificationImpactFlag flags) noexcept;
 
-    bool setTime(long long seconds, unsigned int nseconds, SetTimeFlag flags) noexcept;
+    bool setTime(long long seconds, unsigned int nseconds, enums::domain::SetTimeFlag flags) noexcept;
 
-    bool setUserPassword(gsl::czstring<> user, gsl::czstring<> password, SetUserPasswordFlag flags) noexcept;
+    bool setUserPassword(gsl::czstring<> user, gsl::czstring<> password, enums::domain::SetUserPasswordFlag flags) noexcept;
 
     bool setVcpu(gsl::czstring<> vcpumap, bool state, ModificationImpactFlag flags) noexcept;
 
     bool setVcpus(unsigned int nvcpus) noexcept;
 
-    bool setVcpus(unsigned int nvcpus, VCpuFlag flags) noexcept;
+    bool setVcpus(unsigned int nvcpus, enums::domain::VCpuFlag flags) noexcept;
 
     bool shutdown() noexcept;
 
@@ -699,22 +486,24 @@ class Domain {
 
     // [[nodiscard]] static Domain defineXML();
 };
-}
+} // namespace virt
 #include "virt_wrap/enums/Domain/Domain.hpp"
 namespace virt {
-class Domain::Stats::Record {
+class Domain::StatsRecord {
     friend Connection;
 
     Domain dom;
     std::vector<TypedParameter> params{};
 
-    explicit Record(const virDomainStatsRecord&) noexcept;
+    explicit StatsRecord(const virDomainStatsRecord&) noexcept;
 
   public:
 };
 
-class Domain::StateWReason : public std::variant<StateReason::NoState, StateReason::Running, StateReason::Blocked, StateReason::Paused,
-                                                 StateReason::Shutdown, StateReason::Shutoff, StateReason::Crashed, StateReason::PMSuspended> {
+class Domain::StateWReason
+    : public std::variant<enums::domain::state_reason::NoState, enums::domain::state_reason::Running, enums::domain::state_reason::Blocked,
+                          enums::domain::state_reason::Paused, enums::domain::state_reason::Shutdown, enums::domain::state_reason::Shutoff,
+                          enums::domain::state_reason::Crashed, enums::domain::state_reason::PMSuspended> {
     constexpr State state() const noexcept { return State(EHTag{}, this->index()); }
 };
 
@@ -792,7 +581,7 @@ class Domain::InterfaceView : private virDomainInterface {
 };
 
 struct alignas(alignof(virDomainJobInfo)) Domain::JobInfo : public virDomainJobInfo {
-    [[nodiscard]] constexpr JobType type() const noexcept { return JobType{virDomainJobInfo::type}; }
+    [[nodiscard]] constexpr enums::domain::JobType type() const noexcept { return enums::domain::JobType{virDomainJobInfo::type}; }
 };
 
 // concept Domain::IOThreadInfo
