+++ include/virt_wrap/enums/Domain/SaveRestoreFlag.hpp
@@ -1,11 +1,12 @@
 #pragma once
 #include "virt_wrap/enums/Base.hpp"
+#include "virt_wrap/enums/Domain/Decls.hpp"
 #include "virt_wrap/utility.hpp"
 #include <libvirt/libvirt-domain.h>
 
-namespace virt {
+namespace virt::enums {
 
-class Domain::SaveRestoreFlag : private VirtEnumStorage<unsigned>,
+class domain::SaveRestoreFlag : private VirtEnumStorage<unsigned>,
                                 public VirtCustomEnum,
                                 public VirtEnumBase<SaveRestoreFlag>,
                                 public EnumSetHelper<SaveRestoreFlag> {
@@ -50,9 +51,11 @@ class Domain::SaveRestoreFlag : private VirtEnumStorage<unsigned>,
     constexpr static auto PAUSED = PAUSED_t{};
 
     using VirtEnumBase::VirtEnumBase;
-    constexpr SaveRestoreFlag(unsigned u) noexcept { underlying = u; }
-    constexpr SaveRestoreFlag(virDomainSaveRestoreFlags u) noexcept { underlying = u; }
-    template <typename T, typename = std::enable_if_t<std::is_base_of_v<EBase, T>>> constexpr SaveRestoreFlag(T) noexcept { underlying = T::value; }
+    constexpr explicit SaveRestoreFlag(unsigned u) noexcept { underlying = u; }
+    constexpr explicit SaveRestoreFlag(virDomainSaveRestoreFlags u) noexcept { underlying = u; }
+    template <typename T, typename = std::enable_if_t<std::is_base_of_v<EBase, T>>> constexpr explicit SaveRestoreFlag(T) noexcept {
+        underlying = T::value;
+    }
     template <typename T, typename = std::enable_if_t<std::is_base_of_v<EBase, T>>> constexpr SaveRestoreFlag& operator=(T) noexcept {
         underlying = T::value;
         return *this;
@@ -79,43 +82,49 @@ class Domain::SaveRestoreFlag : private VirtEnumStorage<unsigned>,
 
     constexpr SaveRestoreFlag operator|(SaveRestoreFlag f) const noexcept;
     constexpr SaveRestoreFlag operator|(DEFAULT_t) const noexcept { return *this; }
-    constexpr SaveRestoreFlag operator|(BYPASS_CACHE_t) const noexcept { return virDomainSaveRestoreFlags(underlying | VIR_DOMAIN_SAVE_BYPASS_CACHE); }
-    constexpr SaveRestoreFlag operator|(RUNNING_t) const noexcept { return virDomainSaveRestoreFlags((underlying & ~VIR_DOMAIN_SAVE_PAUSED) | VIR_DOMAIN_SAVE_RUNNING); }
-    constexpr SaveRestoreFlag operator|(PAUSED_t) const noexcept { return virDomainSaveRestoreFlags((underlying & ~VIR_DOMAIN_SAVE_RUNNING) | VIR_DOMAIN_SAVE_PAUSED); }
+    constexpr SaveRestoreFlag operator|(BYPASS_CACHE_t) const noexcept {
+        return virDomainSaveRestoreFlags(underlying | VIR_DOMAIN_SAVE_BYPASS_CACHE);
+    }
+    constexpr SaveRestoreFlag operator|(RUNNING_t) const noexcept {
+        return virDomainSaveRestoreFlags((underlying & ~VIR_DOMAIN_SAVE_PAUSED) | VIR_DOMAIN_SAVE_RUNNING);
+    }
+    constexpr SaveRestoreFlag operator|(PAUSED_t) const noexcept {
+        return virDomainSaveRestoreFlags((underlying & ~VIR_DOMAIN_SAVE_RUNNING) | VIR_DOMAIN_SAVE_PAUSED);
+    }
 
     constexpr static auto from_string(std::string_view sv) { return EnumSetHelper{}.from_string_base(sv); }
 };
 
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::DEFAULT_t, Domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::DEFAULT_t, domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
     return {VIR_DOMAIN_SAVE_BYPASS_CACHE};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::DEFAULT_t, Domain::SaveRestoreFlag::RUNNING_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::DEFAULT_t, domain::SaveRestoreFlag::RUNNING_t) noexcept {
     return {VIR_DOMAIN_SAVE_RUNNING};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::DEFAULT_t, Domain::SaveRestoreFlag::PAUSED_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::DEFAULT_t, domain::SaveRestoreFlag::PAUSED_t) noexcept {
     return {VIR_DOMAIN_SAVE_PAUSED};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::BYPASS_CACHE_t, Domain::SaveRestoreFlag::DEFAULT_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::BYPASS_CACHE_t, domain::SaveRestoreFlag::DEFAULT_t) noexcept {
     return {VIR_DOMAIN_SAVE_BYPASS_CACHE};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::BYPASS_CACHE_t, Domain::SaveRestoreFlag::RUNNING_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::BYPASS_CACHE_t, domain::SaveRestoreFlag::RUNNING_t) noexcept {
     return virDomainSaveRestoreFlags(VIR_DOMAIN_SAVE_BYPASS_CACHE | VIR_DOMAIN_SAVE_RUNNING);
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::BYPASS_CACHE_t, Domain::SaveRestoreFlag::PAUSED_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::BYPASS_CACHE_t, domain::SaveRestoreFlag::PAUSED_t) noexcept {
     return virDomainSaveRestoreFlags(VIR_DOMAIN_SAVE_BYPASS_CACHE | VIR_DOMAIN_SAVE_PAUSED);
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::RUNNING_t, Domain::SaveRestoreFlag::DEFAULT_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::RUNNING_t, domain::SaveRestoreFlag::DEFAULT_t) noexcept {
     return {VIR_DOMAIN_SAVE_RUNNING};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::RUNNING_t, Domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::RUNNING_t, domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
     return virDomainSaveRestoreFlags(VIR_DOMAIN_SAVE_BYPASS_CACHE | VIR_DOMAIN_SAVE_RUNNING);
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::PAUSED_t, Domain::SaveRestoreFlag::DEFAULT_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::PAUSED_t, domain::SaveRestoreFlag::DEFAULT_t) noexcept {
     return {VIR_DOMAIN_SAVE_PAUSED};
 }
-constexpr Domain::SaveRestoreFlag operator|(Domain::SaveRestoreFlag::PAUSED_t, Domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
+constexpr domain::SaveRestoreFlag operator|(domain::SaveRestoreFlag::PAUSED_t, domain::SaveRestoreFlag::BYPASS_CACHE_t) noexcept {
     return virDomainSaveRestoreFlags(VIR_DOMAIN_SAVE_BYPASS_CACHE | VIR_DOMAIN_SAVE_PAUSED);
 }
 
-} // namespace virt
-constexpr unsigned to_integral(virt::Domain::SaveRestoreFlag f) noexcept { return impl_to_integral(f); }
+} // namespace virt::enums
+constexpr unsigned to_integral(virt::domain::SaveRestoreFlag f) noexcept { return impl_to_integral(f); }
